Decompiled func_80079DE8

+++ include/code0/7A430.h
@@ -5,7 +5,7 @@
 
 void func_80079830(void);
 void func_80079C20(s16 spritenum);
-u8 func_80079DE8(u8, u8);
+u8 func_80079DE8(s16, s16);
 void func_8007AD30(void);
 
 #endif
